change references, remove classes unavailable in teigha, change teighga dxf code int to int32

+++ Linq2Acad/Extensions/DbObjectExtensions.cs
@@ -82,7 +82,7 @@ namespace Linq2Acad
         }
         else if (typeof(T) == typeof(long))
         {
-          saveData(getResultBuffer(DxfCode.Int64));
+          saveData(getResultBuffer(DxfCode.Int32));
         }
         else if (typeof(T) == typeof(string))
         {
